move importDataTransfer out of board
